Snapshots without in-situ data
